requirements.txt
nats-py==2.8.0
python-dotenv==1.0.0